(WIP) Moved utils, espresso and schedulers packages under commons.
