+++ app/src/androidTest/java/org/hisp/dhis/android/dataentry/commons/espresso/CircleViewMatchers.java
@@ -1,4 +1,4 @@
-package org.hisp.dhis.android.dataentry.espresso;
+package org.hisp.dhis.android.dataentry.commons.espresso;
 
 import android.support.annotation.NonNull;
 import android.support.test.espresso.matcher.BoundedMatcher;
