+++ app/src/androidTest/java/org/hisp/dhis/android/dataentry/commons/utils/StringUtilsIntegrationTests.java
@@ -1,4 +1,4 @@
-package org.hisp.dhis.android.dataentry.utils;
+package org.hisp.dhis.android.dataentry.commons.utils;
 
 import android.support.test.runner.AndroidJUnit4;
