+++ app/src/androidTest/java/org/hisp/dhis/android/dataentry/reports/ReportsRobot.java
@@ -7,7 +7,7 @@ import static android.support.test.espresso.Espresso.onView;
 import static android.support.test.espresso.assertion.ViewAssertions.matches;
 import static android.support.test.espresso.matcher.ViewMatchers.isAssignableFrom;
 import static org.hamcrest.Matchers.is;
-import static org.hisp.dhis.android.dataentry.espresso.CustomViewMatchers.withToolbarTitle;
+import static org.hisp.dhis.android.dataentry.commons.espresso.CustomViewMatchers.withToolbarTitle;
 
 final class ReportsRobot {
