+++ app/src/main/java/org/hisp/dhis/android/dataentry/commons/database/DbModule.java
@@ -8,7 +8,7 @@ import com.squareup.sqlbrite.SqlBrite;
 
 import org.hisp.dhis.android.core.data.database.DatabaseAdapter;
 import org.hisp.dhis.android.core.data.database.DbOpenHelper;
-import org.hisp.dhis.android.dataentry.utils.SchedulerProvider;
+import org.hisp.dhis.android.dataentry.commons.schedulers.SchedulerProvider;
 
 import javax.inject.Singleton;
