+++ app/src/main/java/org/hisp/dhis/android/dataentry/commons/ui/FontTextInputEditText.java
@@ -9,9 +9,9 @@ import android.support.design.widget.TextInputEditText;
 import android.util.AttributeSet;
 
 import org.hisp.dhis.android.dataentry.R;
-import org.hisp.dhis.android.dataentry.utils.TypefaceManager;
+import org.hisp.dhis.android.dataentry.commons.utils.TypefaceManager;
 
-import static org.hisp.dhis.android.dataentry.utils.Preconditions.isNull;
+import static org.hisp.dhis.android.dataentry.commons.utils.Preconditions.isNull;
 
 public class FontTextInputEditText extends TextInputEditText {
