+++ app/src/main/java/org/hisp/dhis/android/dataentry/commons/utils/TypefaceManager.java
@@ -1,4 +1,4 @@
-package org.hisp.dhis.android.dataentry.utils;
+package org.hisp.dhis.android.dataentry.commons.utils;
 
 import android.content.res.AssetManager;
 import android.graphics.Typeface;
