+++ app/src/main/java/org/hisp/dhis/android/dataentry/launcher/LauncherPresenterImpl.java
@@ -5,7 +5,7 @@ import android.support.annotation.Nullable;
 
 import org.hisp.dhis.android.dataentry.commons.ui.View;
 import org.hisp.dhis.android.dataentry.server.UserManager;
-import org.hisp.dhis.android.dataentry.utils.SchedulerProvider;
+import org.hisp.dhis.android.dataentry.commons.schedulers.SchedulerProvider;
 
 import io.reactivex.disposables.CompositeDisposable;
 import timber.log.Timber;
