+++ app/src/main/java/org/hisp/dhis/android/dataentry/login/LoginPresenterImpl.java
@@ -9,7 +9,7 @@ import org.hisp.dhis.android.dataentry.Components;
 import org.hisp.dhis.android.dataentry.commons.ui.View;
 import org.hisp.dhis.android.dataentry.server.ConfigurationRepository;
 import org.hisp.dhis.android.dataentry.server.UserManager;
-import org.hisp.dhis.android.dataentry.utils.SchedulerProvider;
+import org.hisp.dhis.android.dataentry.commons.schedulers.SchedulerProvider;
 
 import java.io.IOException;
 import java.net.HttpURLConnection;
