+++ app/src/main/java/org/hisp/dhis/android/dataentry/main/MainModule.java
@@ -2,7 +2,7 @@ package org.hisp.dhis.android.dataentry.main;
 
 import org.hisp.dhis.android.dataentry.commons.dagger.PerActivity;
 import org.hisp.dhis.android.dataentry.user.UserRepository;
-import org.hisp.dhis.android.dataentry.utils.SchedulerProvider;
+import org.hisp.dhis.android.dataentry.commons.schedulers.SchedulerProvider;
 
 import dagger.Module;
 import dagger.Provides;
