+++ app/src/main/java/org/hisp/dhis/android/dataentry/reports/ReportsAdapter.java
@@ -23,7 +23,7 @@ import java.util.List;
 import butterknife.BindView;
 import butterknife.ButterKnife;
 
-import static org.hisp.dhis.android.dataentry.utils.StringUtils.htmlify;
+import static org.hisp.dhis.android.dataentry.commons.utils.StringUtils.htmlify;
 
 class ReportsAdapter extends Adapter<ReportsAdapter.ReportViewHolder> {
