+++ app/src/test/java/org/hisp/dhis/android/dataentry/commons/utils/StringUtilsTests.java
@@ -1,4 +1,4 @@
-package org.hisp.dhis.android.dataentry.utils;
+package org.hisp.dhis.android.dataentry.commons.utils;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
