+++ app/src/test/java/org/hisp/dhis/android/dataentry/login/LoginPresenterUnitTests.java
@@ -5,7 +5,7 @@ import org.hisp.dhis.android.core.user.User;
 import org.hisp.dhis.android.dataentry.Components;
 import org.hisp.dhis.android.dataentry.server.ConfigurationRepository;
 import org.hisp.dhis.android.dataentry.server.UserManager;
-import org.hisp.dhis.android.dataentry.utils.MockSchedulersProvider;
+import org.hisp.dhis.android.dataentry.commons.schedulers.MockSchedulersProvider;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
